+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## New
 - Added bar member cocktail reviews
 - Added cocktail sort options: `review_count`, `year`
+- Added ingredients filter options: `distillery`, `origin`
+- Added origins and distilleries to `meta` on GET `/ingredients` endpoint
 
 # v6.4.0
 ## New